Revert "maps + header"
This reverts commit cc611b8af9dc726b6c8ebcf8bc1e326f6bc1608d.

+++ public/assets/js/maps.js
@@ -1,13 +1,12 @@
-
 function initMap(geocoder, resultsMap) {
   var map = new google.maps.Map(document.getElementById('map'), {
     zoom: 8,
-    center: {lat: parseFloat($('#lat').val()), lng: parseFloat($('#lon').val())}
+    center: {lat: parseInt($('#lat').val()), lng: parseInt($('#lon').val())}
        });
   var marker = new google.maps.Marker({
       map: map,
       animation: google.maps.Animation.DROP,
-      position: {lat: parseFloat($('#lat').val()), lng: parseFloat($('#lon').val())}
+      position: {lat: parseInt($('#lat').val()), lng: parseInt($('#lon').val())}
         
   });
   var geocoder = new google.maps.Geocoder();
